backend/requirements.txt
fastapi~=0.115.12
SQLAlchemy~=2.0.40
pydantic~=2.11.4
uvicorn[standard]
asyncpg
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
pydantic[email]